process-images/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
numpy==2.1.1
opencv-python==4.10.0.84
pillow==10.4.0
requests==2.32.3
tk==0.1.0
urllib3==2.2.3
